Add authorization checks to resource actions

+++ app/actions/deleteResource.ts
@@ -2,9 +2,10 @@
 
 import { deleteResource } from '../../lib/services/queries/deleteResource';
 import { revalidatePath } from 'next/cache';
+import { auth } from '~/auth';
 import { deleteCacheResource } from '~/lib/redis/delete-resource';
 import { RESOURCE_CATEGORY } from '~/lib/schemas/category';
-import { isError } from '~/lib/utils/either';
+import { errorResponse, isError } from '~/lib/utils/either';
 
 export const deleteResourceAction = async ({
   resourceId,
@@ -13,6 +14,9 @@ export const deleteResourceAction = async ({
   resourceId: string;
   category: RESOURCE_CATEGORY;
 }) => {
+  const session = await auth();
+  if (!session?.user) return errorResponse('Not authorized');
+
   const response = await deleteResource({ resourceId });
   if (isError(response)) return response;
